+++ OsuDiffCalc/FileProcessor/AnalyzerObjects/SeriesPoint.cs
@@ -1,7 +1,7 @@
 ﻿namespace OsuDiffCalc.FileProcessor.AnalyzerObjects {
 	using System;
 
-	readonly record struct SeriesPoint(float X, float Y) : IComparable<SeriesPoint> {
+	readonly record struct SeriesPoint(double X, double Y) : IComparable<SeriesPoint> {
 		public int CompareTo(SeriesPoint other) => X.CompareTo(other.X);
 	}
 }
